+++ apps/doji-backend/src/entities/Schedule.ts
@@ -1,0 +1,30 @@
+import { ScheduleStatus } from '@libs/api'
+import { Entity, Enum, PrimaryKey, Property } from '@mikro-orm/core'
+
+@Entity()
+export class Schedule {
+  @PrimaryKey()
+  id!: number
+
+  @Property()
+  startTime: Date = new Date()
+
+  @Property()
+  duration: number
+
+  @Property()
+  topic: string
+
+  @Property()
+  fee: number
+
+  @Enum(() => ScheduleStatus)
+  @Property({ default: ScheduleStatus.PENDING })
+  status!: ScheduleStatus
+
+  @Property()
+  coinOnHold: number
+
+  @Property()
+  meetingProviderId: string
+}
