+++ libs/api/src/lib/entities/UserSchedule.ts
@@ -1,0 +1,4 @@
+export interface IUserSchedule {
+  username: string
+  scheduleId: number
+}
